+++ frontend/src/pages/Main/Category/index.tsx
@@ -1,37 +1,12 @@
-import {
-  BackwardArrowSVG,
-  BeanSVG,
-  CafeSVG,
-  CultureSVG,
-  DrinkSVG,
-  ExerciseSVG,
-  ForwardArrowSVG,
-  GameSVG,
-  GuitarSVG,
-  MogackoSVG,
-  SicsaSVG,
-  StudySVG,
-  TravelSVG,
-} from 'assets/svg';
+import { BackwardArrowSVG, BeanSVG, ForwardArrowSVG } from 'assets/svg';
 import useCategory from 'hooks/useCategory';
 import { CategoryType } from 'types/data';
+import { getCategoryIcon } from 'utils/category';
 
 import * as S from './index.styled';
 
 const svgSize = 40;
-const categoryIcon = [
-  { icon: <StudySVG width={svgSize} height={svgSize} /> },
-  { icon: <MogackoSVG width={svgSize} height={svgSize} /> },
-  { icon: <SicsaSVG width={svgSize} height={svgSize} /> },
-  { icon: <CafeSVG width={svgSize} height={svgSize} /> },
-  { icon: <DrinkSVG width={svgSize} height={svgSize} /> },
-  { icon: <ExerciseSVG width={svgSize} height={svgSize} /> },
-  { icon: <GameSVG width={svgSize} height={svgSize} /> },
-  { icon: <TravelSVG width={svgSize} height={svgSize} /> },
-  { icon: <CultureSVG width={svgSize} height={svgSize} /> },
-  { icon: <GuitarSVG width={svgSize} height={svgSize} /> },
-];
-const categorySVG = (id: number) => categoryIcon[id - 1].icon;
+
 interface CategoryProps {
   selectedCategoryId: CategoryType['id'];
   selectCategory: (id: CategoryType['id']) => () => void;
@@ -68,14 +43,13 @@ function Category({
               }
             >
               <S.Button type="button" className={isSelected ? 'select' : ''}>
-                {categorySVG(id)}
+                {getCategoryIcon(id)}
               </S.Button>
               <p>{name}</p>
             </S.ButtonContainer>
           );
         })}
       </S.Box>
-
       <S.FloatingRightButton>
         <ForwardArrowSVG />
       </S.FloatingRightButton>
