+++ frontend/src/pages/Main/WholeGroups/index.tsx
@@ -31,13 +31,11 @@ function WholeGroups({ isFetching, data, refetch, groups }: WholeGroupsProps) {
   return (
     <S.Container>
       {groups.length > 0 ? (
-        <>
-          <S.GroupListBox>
-            {groups.map(group => (
-              <Card group={group} key={group.id} />
-            ))}
-          </S.GroupListBox>
-        </>
+        <S.GroupListBox>
+          {groups.map(group => (
+            <Card group={group} key={group.id} />
+          ))}
+        </S.GroupListBox>
       ) : (
         <NoResult>
           찾고 계신 모임이 없어요 ・゜・(ノД`)
